FSM: Simplified state transition mechanism, removing the concept of StateIdentifiers.
Instead of a State signalling a symbolic "transition request" by
returning a StateIdentifier, a state simply directly constructs its
successor state if it wishes to transition to a new state. This allows
the state to easily pass any context to an appropriate constructor of
the successor, if such a thing is desirable.

Also removed the concept of on_enter or on_exit events, this is better
handled by simply adding another state to the transition chain, that
immediately transitions to the successor.